cap nhat

+++ Eatclean/src/main/java/eatclean/project/demo/repository/LoginRepository.java
@@ -1,17 +1,17 @@
 package eatclean.project.demo.repository;
 
-import java.util.List;
 
+import java.lang.Long;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import eatclean.project.demo.enity.Login;
-import eatclean.project.demo.enity.User;
+
 
 
 
 @Repository
-public interface LoginRepository extends JpaRepository<Login, Long>{
+public interface LoginRepository extends JpaRepository<Login, Integer>{
 	Login findByUsernameAndPassword(String username, String password);
     Login findByUsername(String username);
     Login findById(int id);
